Pause Menu
I made a pause menu by pressing P.
the menu can Resume and quit, but not the other buttons (yet).
If there's any keyboard functionality, wrap it with if(PauseMenu.ifPaused){} or something like that. I left a comment in the PauseMenu function. There should be a build avalible in the Builds folder.