Sprite Billboarding and Even More Buildings
I added a second type of building, which spawns randomly.
I worked on a jumping animation and experimented on putting animations in the project.
I tried to get the sprites to always face the camera (billboarding). At one point it worked. but now it's broken. I figured out it has to do with not getting the camera's angle, which I do not know how 2 fix.